Added action calculation of funds

+++ ServiceMaxTon/Model/DataBase/Cash.cs
@@ -8,7 +8,7 @@ namespace ServiceMaxTon.Model.DataBase
     public class Cash
     {
         public int id { get; set; }
-        public int NetProfit { get; set; }
+        public double NetProfit { get; set; }
         public int DirtyProfit { get; set; }
 
     }
